LayoutTests/platform/ios-simulator/css3/viewport-percentage-lengths/viewport-percentage-lengths-calc-expected.txt

Test for calc support with viewport units.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS style.getPropertyValue("width") is '160px'
PASS style.getPropertyValue("height") is '345px'
PASS style.getPropertyValue("margin-top") is '125px'
PASS style.getPropertyValue("margin-bottom") is '70px'
PASS style.getPropertyValue("margin-left") is '160px'
PASS style.getPropertyValue("margin-right") is '80px'
PASS successfullyParsed is true

TEST COMPLETE
